Move pkg/controller/cloud, pkg/controller/route, pkg/controller/service to k8s.io/cloud-provider
