+++ cmd/cloud-controller-manager/app/apis/config/BUILD
@@ -12,10 +12,10 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/controller/apis/config:go_default_library",
-        "//pkg/controller/service/config:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/service/config:go_default_library",
     ],
 )
