+++ cmd/cloud-controller-manager/app/apis/config/v1alpha1/BUILD
@@ -16,11 +16,11 @@ go_library(
     deps = [
         "//cmd/cloud-controller-manager/app/apis/config:go_default_library",
         "//pkg/controller/apis/config/v1alpha1:go_default_library",
-        "//pkg/controller/service/config/v1alpha1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/conversion:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/service/config/v1alpha1:go_default_library",
         "//staging/src/k8s.io/kube-controller-manager/config/v1alpha1:go_default_library",
     ],
 )
