+++ cmd/cloud-controller-manager/app/options/BUILD
@@ -56,10 +56,10 @@ go_test(
     deps = [
         "//cmd/controller-manager/app/options:go_default_library",
         "//pkg/controller/apis/config:go_default_library",
-        "//pkg/controller/service/config:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/diff:go_default_library",
         "//staging/src/k8s.io/apiserver/pkg/server/options:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/service/config:go_default_library",
         "//staging/src/k8s.io/component-base/config:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
     ],
