+++ cmd/controller-manager/app/options/BUILD
@@ -14,8 +14,8 @@ go_library(
     deps = [
         "//pkg/client/leaderelectionconfig:go_default_library",
         "//pkg/controller/apis/config:go_default_library",
-        "//pkg/controller/service/config:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/service/config:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
         "//staging/src/k8s.io/component-base/config:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
