+++ pkg/controller/BUILD
@@ -112,7 +112,6 @@ filegroup(
         "//pkg/controller/apis/config:all-srcs",
         "//pkg/controller/bootstrap:all-srcs",
         "//pkg/controller/certificates:all-srcs",
-        "//pkg/controller/cloud:all-srcs",
         "//pkg/controller/clusterroleaggregation:all-srcs",
         "//pkg/controller/cronjob:all-srcs",
         "//pkg/controller/daemon:all-srcs",
@@ -131,8 +130,6 @@ filegroup(
         "//pkg/controller/replicaset:all-srcs",
         "//pkg/controller/replication:all-srcs",
         "//pkg/controller/resourcequota:all-srcs",
-        "//pkg/controller/route:all-srcs",
-        "//pkg/controller/service:all-srcs",
         "//pkg/controller/serviceaccount:all-srcs",
         "//pkg/controller/statefulset:all-srcs",
         "//pkg/controller/testutil:all-srcs",
