+++ pkg/controller/apis/config/v1alpha1/BUILD
@@ -30,7 +30,6 @@ go_library(
         "//pkg/controller/replicaset/config/v1alpha1:go_default_library",
         "//pkg/controller/replication/config/v1alpha1:go_default_library",
         "//pkg/controller/resourcequota/config/v1alpha1:go_default_library",
-        "//pkg/controller/service/config/v1alpha1:go_default_library",
         "//pkg/controller/serviceaccount/config/v1alpha1:go_default_library",
         "//pkg/controller/statefulset/config/v1alpha1:go_default_library",
         "//pkg/controller/ttlafterfinished/config/v1alpha1:go_default_library",
@@ -40,6 +39,7 @@ go_library(
         "//staging/src/k8s.io/apimachinery/pkg/conversion:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/service/config/v1alpha1:go_default_library",
         "//staging/src/k8s.io/component-base/config/v1alpha1:go_default_library",
         "//staging/src/k8s.io/kube-controller-manager/config/v1alpha1:go_default_library",
         "//vendor/k8s.io/utils/pointer:go_default_library",
