+++ pkg/controller/apis/config/v1alpha1/doc.go
@@ -32,7 +32,7 @@ limitations under the License.
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/replicaset/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/replication/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/resourcequota/config/v1alpha1
-// +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/service/config/v1alpha1
+// +k8s:conversion-gen=k8s.io/cloud-provider/service/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/serviceaccount/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/statefulset/config/v1alpha1
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/controller/ttlafterfinished/config/v1alpha1
