+++ staging/src/k8s.io/cloud-provider/cloud/BUILD
@@ -6,7 +6,8 @@ go_library(
         "node_controller.go",
         "node_lifecycle_controller.go",
     ],
-    importpath = "k8s.io/kubernetes/pkg/controller/cloud",
+    importpath = "k8s.io/kubernetes/vendor/k8s.io/cloud-provider/cloud",
+    importpath = "k8s.io/cloud-provider/cloud",
     visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/api/core/v1:go_default_library",
