+++ staging/src/k8s.io/cloud-provider/service/config/v1alpha1/BUILD
@@ -10,13 +10,14 @@ go_library(
         "zz_generated.conversion.go",
         "zz_generated.deepcopy.go",
     ],
-    importpath = "k8s.io/kubernetes/pkg/controller/service/config/v1alpha1",
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/cloud-provider/service/config/v1alpha1",
+    importpath = "k8s.io/cloud-provider/service/config/v1alpha1",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/controller/service/config:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/conversion:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
+        "//staging/src/k8s.io/cloud-provider/service/config:go_default_library",
         "//staging/src/k8s.io/kube-controller-manager/config/v1alpha1:go_default_library",
     ],
 )
