+++ staging/src/k8s.io/cloud-provider/service/config/v1alpha1/conversion.go
@@ -18,8 +18,8 @@ package v1alpha1
 
 import (
 	"k8s.io/apimachinery/pkg/conversion"
+	serviceconfig "k8s.io/cloud-provider/service/config"
 	"k8s.io/kube-controller-manager/config/v1alpha1"
-	serviceconfig "k8s.io/kubernetes/pkg/controller/service/config"
 )
 
 // Important! The public back-and-forth conversion functions for the types in this package
