+++ staging/src/k8s.io/cloud-provider/service/config/v1alpha1/zz_generated.conversion.go
@@ -22,10 +22,10 @@ package v1alpha1
 
 import (
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	conversion "k8s.io/apimachinery/pkg/conversion"
-	runtime "k8s.io/apimachinery/pkg/runtime"
-	v1alpha1 "k8s.io/kube-controller-manager/config/v1alpha1"
-	config "k8s.io/kubernetes/pkg/controller/service/config"
+	"k8s.io/apimachinery/pkg/conversion"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/cloud-provider/service/config"
+	"k8s.io/kube-controller-manager/config/v1alpha1"
 )
 
 func init() {
